move some methods from ManuscriptMetadataSynchronizationHandler into ApproveRejectReviewItem
